website: refactor some visuals
Basically, this commit applies the following changes:

(1) Change container-fluid containers into container-xl containers. This
means that the website won't stretch more than 1180px (by which point it
basically looks horrible).

(2) Do away with the previous MD/XL main post view, which had issues
with a lot of images and with long words in titles. We now keep the same
style as LG on the XL dimension; and we've refactored the MD view to
work well with that design.

(3) Display the newsletter in the right spot on the MD view.

(4) Make the newsletter sticky on the LG/XL views.